requirements.txt
Django>=4.2
python-dotenv>=1.0.0
django-allauth>=0.61.0
twilio>=8.10.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
pyotp>=2.9.0
qrcode>=7.4.2
Pillow>=10.0.0
django-cors-headers>=4.3.0